requirements.txt
future==0.18.2
Keras==2.4.3
matplotlib==3.3.4
numpy==1.18.5
opencv-python==4.5.1.48
Pillow==8.0.1
tensorflow==2.3.1

objgraph~=3.5.0